apply filters on searchMedia in SearchRepositoryImpl

+++ data/repositories/search/src/main/java/com/sanaa/search/dataSource/remote/dto/TvShowSearchDto.kt
@@ -18,7 +18,7 @@ data class TvShowSearchDto(
     val releaseDate: String?,
 
     @SerialName("vote_average")
-    val voteAverage: Double?,
+    val voteAverage: Float?,
 
     @SerialName("genre_ids")
     val genreIds: List<Int>?
